Update index.html
Hi! I am a CS undergrad student and I am really intgerested in contributing to open source projects. As this is my first open source contribution I am not sure of how and what to start. Please help me out with this and guide me in the right path for open source.